新增 query privilege 审核单元测试
